+++ app/controllers/user/projects_controller.rb
@@ -10,6 +10,14 @@ class User::ProjectsController < ApplicationController
     pre_fetch_incoming_trusts(@projects)
   end
 
+  def added_by
+    authorize Project, :index?
+    @pager, @projects = pagy(Project.added_by(current_user).not_completed.by_conversion_date.includes(:assigned_to))
+
+    pre_fetch_establishments(@projects)
+    pre_fetch_incoming_trusts(@projects)
+  end
+
   def completed
     authorize Project, :index?
     @pager, @projects = pagy(Project.assigned_to(current_user).completed)