+++ spec/features/users_can_create_voluntary_conversion_projects_spec.rb
@@ -54,7 +54,7 @@ RSpec.feature "Users can create new voluntary conversion projects" do
 
         expect(page).to have_content("You have created a project for #{project.establishment.name}, URN #{project.urn}.")
         expect(page).to have_content("Another person will be assigned to this project.")
-        expect(page).to have_link("Return to project list", href: in_progress_user_projects_path)
+        expect(page).to have_link("View projects you have added", href: added_by_user_projects_path)
       end
 
       it "does not assign the user to the project" do